split the base branch sha guessing logic into a small class hierarchy
