+++ sig/datadog/ci/git/base_branch_sha_detection/branch_metric.rbs
@@ -1,0 +1,20 @@
+module Datadog
+  module CI
+    module Git
+      module BaseBranchShaDetection
+        class BranchMetric
+          attr_reader branch_name: String
+          attr_reader behind: Integer
+          attr_reader ahead: Integer
+          attr_reader base_sha: String
+
+          def initialize: (branch_name: String, behind: Integer, ahead: Integer, base_sha: String) -> void
+
+          def up_to_date?: () -> bool
+
+          def divergence_score: () -> Integer
+        end
+      end
+    end
+  end
+end
