+++ sig/datadog/ci/git/local_repository.rbs
@@ -45,37 +45,11 @@ module Datadog
 
         def self.base_commit_sha: (?base_branch: String?) -> String?
 
-        private
-
-        def self.get_source_branch: () -> String?
-
-        def self.get_remote_name: () -> String
-
         def self.get_upstream_branch: () -> String?
-
-        def self.remove_remote_prefix: (String? branch_name, String remote_name) -> String?
-
-        def self.main_like_branch?: (String? branch_name, String remote_name) -> bool?
-
-        def self.detect_default_branch: (String remote_name) -> String?
-
-        def self.find_fallback_default_branch: (String remote_name) -> String?
-
-        def self.build_candidate_list: (String remote_name) -> Array[String]?
-
-        def self.compute_branch_metrics: (Array[String] candidates, String target_branch) -> Array[BranchMetric]
-
-        def self.find_best_branch: (Array[BranchMetric] metrics, String remote_name) -> String?
-
-        def self.branches_equal?: (String? branch_name, String? default_branch, String remote_name) -> bool
 
         def self.filter_invalid_commits: (Enumerable[String] commits) -> Array[String]
 
         def self.log_failure: (StandardError e, String action) -> void
-
-        def self.check_and_fetch_base_branches: (Array[String] branches, String remote_name) -> void
-
-        def self.check_and_fetch_branch: (String branch, String remote_name) -> void
       end
     end
   end
